add CBA_fnc_switchableAttachments
