+++ addons/accessory/CfgFunctions.hpp
@@ -1,0 +1,7 @@
+class CfgFunctions {
+    class CBA {
+        class Inventory {
+            PATHTO_FNC(switchableAttachments);
+        };
+    };
+};
