+++ addons/accessory/config.cpp
@@ -14,14 +14,5 @@ class CfgPatches {
     };
 };
 
-class Extended_PreStart_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_FILE(XEH_preStart));
-    };
-};
-
-class Extended_PreInit_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_FILE(XEH_preInit));
-    };
-};
+#include "CfgEventHandlers.hpp"
+#include "CfgFunctions.hpp"
